Refactor onboarding movies step logic
- Extract the suggestion-pool engine (constants + fetchSuggestionPool/fetchGlobalPool/
  shapeFilm/mergeUnique/scoreCandidates/dedupByCollection) into suggestionPool.js,
  byte-identical scoring values and Supabase query filters
- Add useSuggestionPool (mount-once fetch, preserved) and useMovieSearch (300ms
  debounce, popularity sort, top-8 — moved verbatim)
- Extract MovieCard / CardSkeletonRow / SearchDropdown presentational pieces
- MoviesStep.jsx is now a thin composition layer; render is byte-identical except
  the clear-button handler and the dropdown component (both behavior-identical)
- Add focused pure-function tests for the engine (no live Supabase/TMDB)
- No UI/copy/layout/styling change; auth/routing/Supabase writes/localStorage/
  validation/onboarding completion untouched
